==> Colab Notebooks/Step1_Life expectancy cleaning.ipynb <==
import numpy as np
import pandas as pd

from google.colab import drive
drive.mount("/content/gdrive")

lfd_df = pd.read_csv("/content/gdrive/MyDrive/Raw Dataset/Life Expectancy Data.csv")

lfd_df.shape

lfd_df.Country.unique()

lfd_df.head(16)

lfd_df.columns

lfd_df.info()

col_names = []
for i in range(len(lfd_df.columns)):
    col_names.append(lfd_df.columns[i].strip())

lfd_df.columns = col_names

lfd_df.columns

cols_to_be_removed = ['Hepatitis B', 'Measles', 'Polio', 'Diphtheria']

lfd_df.drop(cols_to_be_removed, axis=1, inplace=True)

lfd_df.columns

lfd_df[lfd_df[['HIV/AIDS', 'thinness  1-19 years', 'thinness 5-9 years']].isna().any(axis=1)].shape

lfd_df.shape

#### Required fields for this dataset
- Status (Developed or Developing status)
- Adult Mortality (Adult Mortality Rates of both sexes (probability of dying between 15 and 60 years per 1000 population))
- infant deaths (Number of Infant Deaths per 1000 population)
- Alcohol (recorded per capita (15+) consumption (in liters of pure alcohol))
- percentage expenditure (Expenditure on health as a percentage of Gross Domestic Product per capita(%))
- Hepatitis B (Hepatitis B (HepB) immunization coverage among 1-year-olds (%))
- Measles (Measles - number of reported cases per 1000 population)
- under-five deaths (Number of under-five deaths per 1000 population)
- Polio (Polio (Pol3) immunization coverage among 1-year-olds (%))
- Total expenditure (General government expenditure on health as a percentage of total government expenditure (%))
- Diphtheria (Diphtheria tetanus toxoid and pertussis (DTP3) immunization coverage among 1-year-olds (%))
- HIV/AIDS (Deaths per 1 000 live births HIV/AIDS (0-4 years))
- GDP (Gross Domestic Product per capita (in USD))
- Income composition of resources (Human Development Index in terms of income composition of resources (index ranging from 0 to 1))
- Schooling (Number of years of Schooling(years))


required_columns = ['Country', 'Year', 'Status', 'Life expectancy', 'Adult Mortality', 'infant deaths' , 'Alcohol', 'percentage expenditure', 'under-five deaths', 'Total expenditure', 'HIV/AIDS', 'thinness  1-19 years', 'Income composition of resources','Schooling']
ldf_df2 = lfd_df[required_columns]
ldf_df2.sort_values(by=['Country','Year'], ignore_index=True,inplace=True)
ldf_df2

ldf_df2.columns

ldf_df2.Status.unique()

ldf_df2.to_csv("life_expectancy_dataset_cleaned.csv")
