==> Colab Notebooks/Step2_UN gdp cleaning.ipynb <==
import pandas as pd
import numpy as np

from google.colab import drive
drive.mount("/content/gdrive")

gdp_df = pd.read_csv('/content/gdrive/MyDrive/Raw Dataset/gdp.csv')
gdp_df.columns

gdp_df = gdp_df.melt(id_vars=["Country Name"], 
        var_name="year", 
        value_name="Gdp")

gdp_df.columns = ['Country', 'Year', 'GDP']
gdp_df

gdp_df.Year = pd.to_numeric(gdp_df.Year)

gdp_df = gdp_df[gdp_df.Year<2016]

gdp_df = gdp_df.sort_values(by=['Country', 'Year'],ignore_index=True)
gdp_df

gdp_df.isna().sum()

gdp_df.to_csv("GDP_data.csv")
