==> Colab Notebooks/Step4_WHO BMI Cleaning.ipynb <==
from google.colab import drive
drive.mount("/content/gdrive")

import pandas as pd

df = pd.read_csv('/content/gdrive/MyDrive/Raw Dataset/bmi.csv')

df.head()

df.columns

df.Dim2.unique()

required_columns = ['Location', 'Period', 'Dim1', 'Dim2', 'FactValueNumeric']
df2 = df[required_columns]
df2.columns = ['Country', 'Year', 'Gender', 'Age_group', 'bmi']

df2 = df2[df2.Year>1999]
df2 = df2[df2.Year<2016]

df2 = df2.sort_values(by=['Country', 'Year'],ignore_index=True)
df2

df2.Age_group.unique()

df2 = df2[df2.Age_group!='5-19  years']
df2 = df2[df2.Age_group!='18+  years']

df2.Age_group.unique()

df2.sort_values(by=['Country','Year'],ignore_index=True, inplace=True)

df2.head(20)

df3 = pd.DataFrame(df2.groupby(['Country','Year','Gender'])['bmi'].mean())
type(df3)
# df3

df3

male_data = df3.query('Gender == "Male"')
male_data = male_data.sort_values(by=['Country', 'Year'],ignore_index=True)
male_data


female_data = df3.query('Gender == "Female"')
female_data= female_data.sort_values(by=['Country', 'Year'],ignore_index=True)

female_data



both_data = df3.query('Gender == "Both sexes"')
both_data= both_data.sort_values(by=['Country', 'Year'],ignore_index=True)

both_data

df4 = df2[['Country','Year']]
df4.drop_duplicates(inplace=True)

df4.shape

df4.sort_values(by=['Country','Year'],ignore_index=True, inplace=True)

df4['Male_bmi'] = male_data
df4['Female_bmi'] = female_data
df4['bmi'] = both_data
df4

df4.info()

df4.isna().sum()

df4.to_csv("bmi_genderwise.csv")

ENd of code
