==> Colab Notebooks/Step5_WHO homicide cleaning.ipynb <==
from google.colab import drive
drive.mount("/content/gdrive")

import pandas as pd

homicide_df = pd.read_csv('/content/gdrive/MyDrive/Raw Dataset/homicides per 100000.csv')
homicide_df

homicide_df.columns

required_columns = ['Location', 'Period', 'Dim1', 'FactValueNumeric']
homicide_df2 = homicide_df[required_columns]
homicide_df2.columns= ['Country', 'Year', 'Gender', 'Homicides']

homicide_df2 = homicide_df2.sort_values(by=['Country', 'Year'],ignore_index=True)
homicide_df2

homicide_df2 = homicide_df2[homicide_df2.Year<2016]

homicide_df3= homicide_df2[['Country', 'Year']]
homicide_df3.drop_duplicates(inplace=True)
homicide_df3 = homicide_df3.sort_values(by=['Country', 'Year'],ignore_index=True)
homicide_df3

male_data = homicide_df2.query('Gender == "Male"')
male_data = male_data.sort_values(by=['Country', 'Year'],ignore_index=True)
male_data

female_data = homicide_df2.query('Gender == "Female"')
female_data= female_data.sort_values(by=['Country', 'Year'],ignore_index=True)
female_data

both_data = homicide_df2.query('Gender == "Both sexes"')
both_data= both_data.sort_values(by=['Country', 'Year'],ignore_index=True)

both_data

homicide_df3['Male_homicides'] = male_data['Homicides']
homicide_df3['Female_homicides'] = female_data['Homicides']
homicide_df3['Homicides'] = both_data['Homicides']
homicide_df3

homicide_df3.to_csv("homicide_data_genderwise.csv")
