==> Colab Notebooks/Step6_WHO sanitation cleaning.ipynb <==
from google.colab import drive
drive.mount("/content/gdrive")

import pandas as pd

sanitation_df = pd.read_csv('/content/gdrive/MyDrive/Raw Dataset/basic sanitation.csv')
sanitation_df

sanitation_df.columns

required_columns = ['Location', 'Period', 'Dim1' , 'Value']
sanitation_df2 = sanitation_df[required_columns]

sanitation_df2.columns= ['Country', 'Year', 'Area', 'Sanitation']
sanitation_df2



sanitation_df2[sanitation_df2.Country=="Canada"]
sanitation_df2 = sanitation_df2[sanitation_df2.Year<2016]

sanitation_df2 = sanitation_df2[sanitation_df2.Area=="Total"]

sanitation_df2.drop('Area', axis='columns', inplace=True)

sanitation_df2.shape

sanitation_df2.sort_values(by=['Country','Year'],ignore_index=True, inplace=True)

sanitation_df2.head(20)

sanitation_df3 = sanitation_df2.copy()
sanitation_df3.drop_duplicates(inplace=True)
sanitation_df3 = sanitation_df3.sort_values(by=['Country', 'Year'],ignore_index=True)
sanitation_df3

sanitation_df3.to_csv("sanitation_data.csv")
