==> Colab Notebooks/Step7_Data_Merge.ipynb <==
import numpy as np
import pandas as pd

from google.colab import drive
drive.mount("/content/gdrive")

Loading life expectancy data

lfd_df = pd.read_csv("/content/gdrive/MyDrive/Cleaned Dataset/life_expectancy_dataset_cleaned.csv")

lfd_df.columns

lfd_df.drop('Unnamed: 0', axis=1, inplace=True)

lfd_df.rename(columns={'thinness  1-19 years':'thinness'}, inplace=True)

lfd_df.columns

Loading population data

pop_df=pd.read_csv("/content/gdrive/MyDrive/Cleaned Dataset/population_data.csv")

pop_df.columns

pop_df.drop('Unnamed: 0', axis=1, inplace=True)
pop_df.columns

Loading gdp data

gdp_df=pd.read_csv("/content/gdrive/MyDrive/Cleaned Dataset/GDP_data.csv")

gdp_df.columns

gdp_df.drop('Unnamed: 0', axis=1, inplace=True)
gdp_df.columns

#method to deal with country with miltiple names

# !pip install fuzzymatcher
# from fuzzymatcher import link_table, fuzzy_left_join

# merge_df = fuzzy_left_join(lfd_df[['Country']], gdp_df[['Country']], ["Country"], ["Country"])
# merge_df = merge_df[["Country_left"]]

# merge_df.drop_duplicates(inplace=True)
# merge_df.sort_values(by=['Country_left'],ignore_index=True, inplace=True)

# a = merge_df.Country_left.unique().tolist()
# gdp_df[~gdp_df.Country.isin(a)].Country

new_list=[]
for country in lfd_df.Country.tolist():
    if country not in gdp_df.Country.tolist():
        new_list.append(country)

set1 = set(new_list)

new_list=[]
for country in lfd_df.Country.tolist():
    if country not in pop_df.Country.tolist():
        new_list.append(country)

set2 = set(new_list)
print(set1 == set2)


set1

new_list=[]
for country in pop_df.Country.tolist():
    if country not in lfd_df.Country.tolist():
        new_list.append(country)

set3 = set(new_list)
set3

# lfd_df[lfd_df.Country=='Cook Islands']
# lfd_df[lfd_df.Country=='Niue']

dict1={
    'Bahamas, The':'Bahamas',
    'Congo, Rep.':'Congo',

    "Cote d'Ivoire":"Côte d'Ivoire",
    "Korea, Dem. People's Rep.": "Democratic People's Republic of Korea",
    'Congo, Dem. Rep.':'Democratic Republic of the Congo',
    'Egypt, Arab Rep.':'Egypt',
    'Gambia, The':'Gambia',
    'Iran, Islamic Rep.':'Iran',
    'Kyrgyz Republic':'Kyrgyzstan',
    'Lao PDR':"Lao People's Democratic Republic",
    'Micronesia, Fed. Sts.':'Micronesia',

    'Korea, Rep.':'Republic of Korea',
    'Moldova':'Republic of Moldova',
    'St. Kitts and Nevis':'Saint Kitts and Nevis',
    'St. Lucia':'Saint Lucia',
    'St. Vincent and the Grenadines':'Saint Vincent and the Grenadines',
    'Slovak Republic':'Slovakia',
    'Eswatini':'Swaziland',
    'Turkiye':'Turkey',
    'United States':'United States of America',
    'Venezuela, RB':'Venezuela',
    'Yemen, Rep.':'Yemen'
    }
gdp_df["Country"].replace(dict1, inplace=True)
pop_df["Country"].replace(dict1, inplace=True)

dict2={'The former Yugoslav republic of Macedonia':'North Macedonia',
       'United Republic of Tanzania':'Tanzania',
       'United Kingdom of Great Britain and Northern Ireland':'United Kingdom'
       }
lfd_df["Country"].replace(dict2, inplace=True)

new_list=[]
for country in gdp_df.Country.tolist():
    if country not in lfd_df.Country.tolist():
        new_list.append(country)
setd2 = set(new_list)

new_list=[]
for country in lfd_df.Country.tolist():
    if country not in gdp_df.Country.tolist():
        new_list.append(country)

setd1 = set(new_list)

setd1, setd2

print('lfd',lfd_df.Country.nunique(),'\ngdp', gdp_df.Country.nunique(), '\npop', pop_df.Country.nunique())

pop_df = pop_df[pop_df.Country.apply(lambda x:x not in list(setd2))].sort_values(by=['Country'],ignore_index=True)
gdp_df = gdp_df[gdp_df.Country.apply(lambda x:x not in list(setd2))].sort_values(by=['Country'],ignore_index=True)

# d = pd.concat([merge_df, lfd_df[["Country"]].drop_duplicates().sort_values(by=['Country'],ignore_index=True)], axis=1)
# d[d.Country_left!=d.Country]

Loading suicide data

suicide_df=pd.read_csv("/content/gdrive/MyDrive/Cleaned Dataset/Suicide_per_100000.csv")

suicide_df.columns

suicide_df.drop('Unnamed: 0', axis=1, inplace=True)
suicide_df.columns

Loading homicide data

homicide_df=pd.read_csv("/content/gdrive/MyDrive/Cleaned Dataset/homicide_data_genderwise.csv")

homicide_df.drop('Unnamed: 0', axis=1, inplace=True)
homicide_df.columns

homicide_df.columns

new_list=[]
for country in lfd_df.Country.tolist():
    if country not in homicide_df.Country.tolist():
        new_list.append(country)

set4 = set(new_list)

new_list=[]
for country in lfd_df.Country.tolist():
    if country not in suicide_df.Country.tolist():
        new_list.append(country)

set5 = set(new_list)
print(set4==set5)

set4

new_list=[]
for country in suicide_df.Country.tolist():
    if country not in lfd_df.Country.tolist():
        new_list.append(country)

set6 = set(new_list)
set6

dict3={
    'Bolivia (Plurinational State of)':'Bolivia',
    'Côte d’Ivoire':"Côte d'Ivoire",
    'Iran (Islamic Republic of)':'Iran',
    'Micronesia (Federated States of)':'Micronesia',
    'The former Yugoslav Republic of Macedonia': 'North Macedonia',
    'Eswatini':'Swaziland',
    'Türkiye':'Turkey',
    'United Kingdom of Great Britain and Northern Ireland':'United Kingdom',
    'United Republic of Tanzania': 'Tanzania',
    'Venezuela (Bolivarian Republic of)':'Venezuela',
    'Viet Nam':'Vietnam'
    }
homicide_df["Country"].replace(dict3, inplace=True)
suicide_df["Country"].replace(dict3, inplace=True)

new_list=[]
for country in lfd_df.Country.tolist():
    if country not in homicide_df.Country.tolist():
        new_list.append(country)
setd3 = set(new_list)

new_list=[]
for country in homicide_df.Country.tolist():
    if country not in lfd_df.Country.tolist():
        new_list.append(country)

setd4 = set(new_list)

setd3, setd4

lfd_df[lfd_df.Country.apply(lambda x:x in list(setd4))]

print('lfd',lfd_df.Country.nunique(),'\nhomicide', homicide_df.Country.nunique(), '\nsuicide', suicide_df.Country.nunique())

Loading bmi data

bmi_df=pd.read_csv("/content/gdrive/MyDrive/Cleaned Dataset/bmi_genderwise.csv")

bmi_df.columns

bmi_df.drop('Unnamed: 0', axis=1, inplace=True)
bmi_df.columns

new_list=[]
for country in lfd_df.Country.tolist():
    if country not in bmi_df.Country.tolist():
        new_list.append(country)

set7 = set(new_list)
set7

new_list=[]
for country in bmi_df.Country.tolist():
    if country not in lfd_df.Country.tolist():
        new_list.append(country)

set8 = set(new_list)
set8

dict4={
    'Bolivia (Plurinational State of)':'Bolivia',
    'Côte d’Ivoire':"Côte d'Ivoire",
    'Iran (Islamic Republic of)':'Iran',
    'Micronesia (Federated States of)':'Micronesia',
    'The former Yugoslav Republic of Macedonia': 'North Macedonia',
    'Eswatini':'Swaziland',
    'Türkiye':'Turkey',
    'United Kingdom of Great Britain and Northern Ireland':'United Kingdom',
    'United Republic of Tanzania': 'Tanzania',
    'Venezuela (Bolivarian Republic of)':'Venezuela',
    'Viet Nam':'Vietnam'
    }
bmi_df["Country"].replace(dict4, inplace=True)

new_list=[]
for country in lfd_df.Country.tolist():
    if country not in bmi_df.Country.tolist():
        new_list.append(country)

setd5 = set(new_list)

new_list=[]
for country in bmi_df.Country.tolist():
    if country not in lfd_df.Country.tolist():
        new_list.append(country)

setd6 = set(new_list)

setd5, setd6

print('lfd',lfd_df.Country.nunique(),'\nbmi', bmi_df.Country.nunique())

bmi_df = bmi_df[bmi_df.Country.apply(lambda x:x not in list(setd6))].sort_values(by=['Country'],ignore_index=True)

Loading sanitation data

sani_df=pd.read_csv("/content/gdrive/MyDrive/Cleaned Dataset/sanitation_data.csv")

sani_df.columns

sani_df.drop('Unnamed: 0', axis=1, inplace=True)
sani_df.columns

new_list=[]
for country in lfd_df.Country.tolist():
    if country not in sani_df.Country.tolist():
        new_list.append(country)

set9 = set(new_list)
set9

new_list=[]
for country in sani_df.Country.tolist():
    if country not in lfd_df.Country.tolist():
        new_list.append(country)

set10 = set(new_list)
set10

dict5={
    'Bolivia (Plurinational State of)':'Bolivia',
    'Côte d’Ivoire':"Côte d'Ivoire",
    'Iran (Islamic Republic of)':'Iran',
    'Micronesia (Federated States of)':'Micronesia',
    'The former Yugoslav Republic of Macedonia': 'North Macedonia',
    'Eswatini':'Swaziland',
    'Türkiye':'Turkey',
    'United Kingdom of Great Britain and Northern Ireland':'United Kingdom',
    'United Republic of Tanzania': 'Tanzania',
    'Venezuela (Bolivarian Republic of)':'Venezuela',
    'Viet Nam':'Vietnam'
    }
sani_df["Country"].replace(dict5, inplace=True)

new_list=[]
for country in lfd_df.Country.tolist():
    if country not in sani_df.Country.tolist():
        new_list.append(country)

setd7 = set(new_list)

new_list=[]
for country in sani_df.Country.tolist():
    if country not in lfd_df.Country.tolist():
        new_list.append(country)

setd8 = set(new_list)

setd7, setd8

print('lfd',lfd_df.Country.nunique(),'\nsani', sani_df.Country.nunique())

sani_df = sani_df[sani_df.Country.apply(lambda x:x not in list(setd8))].sort_values(by=['Country'],ignore_index=True)

# Merging all data

print('Shape of dataframe')
print('lfd',lfd_df.shape,'\ngdp', gdp_df.shape, '\npop',pop_df.shape,'\nsani', sani_df.shape,'\nsuiside', suicide_df.shape,'\nhomicide', homicide_df.shape,'\nbmi', bmi_df.shape)

print('Number of unique countries in each dataframe')
print('lfd',lfd_df.Country.nunique(),'\ngdp', gdp_df.Country.nunique(), '\npop',pop_df.Country.nunique(),'\nsani', sani_df.Country.nunique(),'\nsuiside', suicide_df.Country.nunique(),'\nhomicide', homicide_df.Country.nunique(),'\nbmi', bmi_df.Country.nunique())

df1=pd.merge(lfd_df,suicide_df, left_on=['Country','Year'], right_on=['Country','Year'], how='outer')

df1.shape

df1.info()

df2=pd.merge(df1,homicide_df, left_on=['Country','Year'], right_on=['Country','Year'], how='outer')

df2.shape

df2.info()

df3=pd.merge(df2,bmi_df, left_on=['Country','Year'], right_on=['Country','Year'], how='outer')

df3.shape

df3.info()

df4=pd.merge(df3,sani_df, left_on=['Country','Year'], right_on=['Country','Year'], how='outer')

df4.shape

df4.info()

df5=pd.merge(df4,gdp_df, left_on=['Country','Year'], right_on=['Country','Year'], how='outer')

df5.shape

df5.info()

df6=pd.merge(df5,pop_df, left_on=['Country','Year'], right_on=['Country','Year'], how='outer')

df6.shape

df6.Country.nunique()

df6.info()

df6.isna().sum()

df7 = df6.dropna(subset=['Life expectancy','Adult Mortality', 'percentage expenditure'], how='all')

df7.shape

df7.sort_values(by=['Country','Year'], ignore_index=True,inplace=True)

df7.to_csv("Merged_data.csv")
