==> Colab Notebooks/Step8_creating required data.ipynb <==
from google.colab import drive
drive.mount("/content/gdrive")

import pandas as pd
import numpy as np

czechia_data = [    {"Country": "Czechia", "Year": 2000, "Avg. Years of Schooling": 11.3, "HDI (Income Composition of Resources)": 0.802},    {"Country": "Czechia", "Year": 2001, "Avg. Years of Schooling": 11.4, "HDI (Income Composition of Resources)": 0.812},    {"Country": "Czechia", "Year": 2002, "Avg. Years of Schooling": 11.6, "HDI (Income Composition of Resources)": 0.828},    {"Country": "Czechia", "Year": 2003, "Avg. Years of Schooling": 11.8, "HDI (Income Composition of Resources)": 0.845},    {"Country": "Czechia", "Year": 2004, "Avg. Years of Schooling": 12.0, "HDI (Income Composition of Resources)": 0.861},    {"Country": "Czechia", "Year": 2005, "Avg. Years of Schooling": 12.2, "HDI (Income Composition of Resources)": 0.874},    {"Country": "Czechia", "Year": 2006, "Avg. Years of Schooling": 12.4, "HDI (Income Composition of Resources)": 0.887},    {"Country": "Czechia", "Year": 2007, "Avg. Years of Schooling": 12.6, "HDI (Income Composition of Resources)": 0.899},    {"Country": "Czechia", "Year": 2008, "Avg. Years of Schooling": 12.8, "HDI (Income Composition of Resources)": 0.910},    {"Country": "Czechia", "Year": 2009, "Avg. Years of Schooling": 13.0, "HDI (Income Composition of Resources)": 0.919},    {"Country": "Czechia", "Year": 2010, "Avg. Years of Schooling": 13.2, "HDI (Income Composition of Resources)": 0.926},    {"Country": "Czechia", "Year": 2011, "Avg. Years of Schooling": 13.4, "HDI (Income Composition of Resources)": 0.931},    {"Country": "Czechia", "Year": 2012, "Avg. Years of Schooling": 13.5, "HDI (Income Composition of Resources)": 0.935},    {"Country": "Czechia", "Year": 2013, "Avg. Years of Schooling": 13.7, "HDI (Income Composition of Resources)": 0.938},    {"Country": "Czechia", "Year": 2014, "Avg. Years of Schooling": 13.8, "HDI (Income Composition of Resources)": 0.940},    {"Country": "Czechia", "Year": 2015, "Avg. Years of Schooling": 13.9, "HDI (Income Composition of Resources)": 0.942}]
Côte_dIvoire_data = [    {"Country": "Côte d'Ivoire", "Year": 2000, "Avg. Years of Schooling": 3.2, "HDI (Income Composition of Resources)": 0.361},    {"Country": "Côte d'Ivoire", "Year": 2001, "Avg. Years of Schooling": 3.3, "HDI (Income Composition of Resources)": 0.363},    {"Country": "Côte d'Ivoire", "Year": 2002, "Avg. Years of Schooling": 3.3, "HDI (Income Composition of Resources)": 0.366},    {"Country": "Côte d'Ivoire", "Year": 2003, "Avg. Years of Schooling": 3.4, "HDI (Income Composition of Resources)": 0.370},    {"Country": "Côte d'Ivoire", "Year": 2004, "Avg. Years of Schooling": 3.5, "HDI (Income Composition of Resources)": 0.375},    {"Country": "Côte d'Ivoire", "Year": 2005, "Avg. Years of Schooling": 3.5, "HDI (Income Composition of Resources)": 0.380},    {"Country": "Côte d'Ivoire", "Year": 2006, "Avg. Years of Schooling": 3.6, "HDI (Income Composition of Resources)": 0.386},    {"Country": "Côte d'Ivoire", "Year": 2007, "Avg. Years of Schooling": 3.7, "HDI (Income Composition of Resources)": 0.392},    {"Country": "Côte d'Ivoire", "Year": 2008, "Avg. Years of Schooling": 3.8, "HDI (Income Composition of Resources)": 0.399},    {"Country": "Côte d'Ivoire", "Year": 2009, "Avg. Years of Schooling": 3.9, "HDI (Income Composition of Resources)": 0.406},    {"Country": "Côte d'Ivoire", "Year": 2010, "Avg. Years of Schooling": 4.0, "HDI (Income Composition of Resources)": 0.414},    {"Country": "Côte d'Ivoire", "Year": 2011, "Avg. Years of Schooling": 4.1, "HDI (Income Composition of Resources)": 0.422},    {"Country": "Côte d'Ivoire", "Year": 2012, "Avg. Years of Schooling": 4.2, "HDI (Income Composition of Resources)": 0.429},    {"Country": "Côte d'Ivoire", "Year": 2013, "Avg. Years of Schooling": 4.3, "HDI (Income Composition of Resources)": 0.436},    {"Country": "Côte d'Ivoire", "Year": 2014, "Avg. Years of Schooling": 4.4, "HDI (Income Composition of Resources)": 0.442},    {"Country": "Côte d'Ivoire", "Year": 2015, "Avg. Years of Schooling": 4.5, "HDI (Income Composition of Resources)": 0.448},]
congo_data = [    {"Country": "Democratic Republic of the Congo", "Year": 2000, "Avg. Years of Schooling": 4.5, "HDI (Income Composition of Resources)": 0.285},    {"Country": "Democratic Republic of the Congo", "Year": 2001, "Avg. Years of Schooling": 4.4, "HDI (Income Composition of Resources)": 0.284},    {"Country": "Democratic Republic of the Congo", "Year": 2002, "Avg. Years of Schooling": 4.4, "HDI (Income Composition of Resources)": 0.287},    {"Country": "Democratic Republic of the Congo", "Year": 2003, "Avg. Years of Schooling": 4.4, "HDI (Income Composition of Resources)": 0.290},    {"Country": "Democratic Republic of the Congo", "Year": 2004, "Avg. Years of Schooling": 4.4, "HDI (Income Composition of Resources)": 0.293},    {"Country": "Democratic Republic of the Congo", "Year": 2005, "Avg. Years of Schooling": 4.5, "HDI (Income Composition of Resources)": 0.297},    {"Country": "Democratic Republic of the Congo", "Year": 2006, "Avg. Years of Schooling": 4.5, "HDI (Income Composition of Resources)": 0.300},    {"Country": "Democratic Republic of the Congo", "Year": 2007, "Avg. Years of Schooling": 4.6, "HDI (Income Composition of Resources)": 0.303},    {"Country": "Democratic Republic of the Congo", "Year": 2008, "Avg. Years of Schooling": 4.6, "HDI (Income Composition of Resources)": 0.306},    {"Country": "Democratic Republic of the Congo", "Year": 2009, "Avg. Years of Schooling": 4.7, "HDI (Income Composition of Resources)": 0.310},    {"Country": "Democratic Republic of the Congo", "Year": 2010, "Avg. Years of Schooling": 4.8, "HDI (Income Composition of Resources)": 0.314},    {"Country": "Democratic Republic of the Congo", "Year": 2011, "Avg. Years of Schooling": 4.9, "HDI (Income Composition of Resources)": 0.318},    {"Country": "Democratic Republic of the Congo", "Year": 2012, "Avg. Years of Schooling": 5.0, "HDI (Income Composition of Resources)": 0.322},    {"Country": "Democratic Republic of the Congo", "Year": 2013, "Avg. Years of Schooling": 5.1, "HDI (Income Composition of Resources)": 0.326},    {"Country": "Democratic Republic of the Congo", "Year": 2014, "Avg. Years of Schooling": 5.2, "HDI (Income Composition of Resources)": 0.329},    {"Country": "Democratic Republic of the Congo", "Year": 2015, "Avg. Years of Schooling": 5.3, "HDI (Income Composition of Resources)": 0.333}]
south_korea_data = [{'Country': 'Republic of Korea', 'Year': 2000, 'Avg. Years of Schooling': 9.8, 'HDI (Income Composition of Resources)': 0.682},        {'Country': 'Republic of Korea', 'Year': 2001, 'Avg. Years of Schooling': 9.9, 'HDI (Income Composition of Resources)': 0.696},        {'Country': 'Republic of Korea', 'Year': 2002, 'Avg. Years of Schooling': 10.0, 'HDI (Income Composition of Resources)': 0.710},        {'Country': 'Republic of Korea', 'Year': 2003, 'Avg. Years of Schooling': 10.2, 'HDI (Income Composition of Resources)': 0.719},        {'Country': 'Republic of Korea', 'Year': 2004, 'Avg. Years of Schooling': 10.3, 'HDI (Income Composition of Resources)': 0.732},        {'Country': 'Republic of Korea', 'Year': 2005, 'Avg. Years of Schooling': 10.4, 'HDI (Income Composition of Resources)': 0.745},        {'Country': 'Republic of Korea', 'Year': 2006, 'Avg. Years of Schooling': 10.6, 'HDI (Income Composition of Resources)': 0.760},        {'Country': 'Republic of Korea', 'Year': 2007, 'Avg. Years of Schooling': 10.7, 'HDI (Income Composition of Resources)': 0.773},        {'Country': 'Republic of Korea', 'Year': 2008, 'Avg. Years of Schooling': 10.8, 'HDI (Income Composition of Resources)': 0.784},        {'Country': 'Republic of Korea', 'Year': 2009, 'Avg. Years of Schooling': 10.9, 'HDI (Income Composition of Resources)': 0.795},        {'Country': 'Republic of Korea', 'Year': 2010, 'Avg. Years of Schooling': 11.0, 'HDI (Income Composition of Resources)': 0.805},        {'Country': 'Republic of Korea', 'Year': 2011, 'Avg. Years of Schooling': 11.2, 'HDI (Income Composition of Resources)': 0.816},        {'Country': 'Republic of Korea', 'Year': 2012, 'Avg. Years of Schooling': 11.3, 'HDI (Income Composition of Resources)': 0.825},        {'Country': 'Republic of Korea', 'Year': 2013, 'Avg. Years of Schooling': 11.4, 'HDI (Income Composition of Resources)': 0.833},        {'Country': 'Republic of Korea', 'Year': 2014, 'Avg. Years of Schooling': 11.5, 'HDI (Income Composition of Resources)': 0.839},        {'Country': 'Republic of Korea', 'Year': 2015, 'Avg. Years of Schooling': 11.6, 'HDI (Income Composition of Resources)': 0.846}]
Moldova_data = [
    {"Country": "Republic of Moldova", "Year": 2000, "Avg. Years of Schooling": 10.2, "HDI (Income Composition of Resources)": 0.660},
    {"Country": "Republic of Moldova", "Year": 2001, "Avg. Years of Schooling": 10.3, "HDI (Income Composition of Resources)": 0.660},
    {"Country": "Republic of Moldova", "Year": 2002, "Avg. Years of Schooling": 10.3, "HDI (Income Composition of Resources)": 0.670},
    {"Country": "Republic of Moldova", "Year": 2003, "Avg. Years of Schooling": 10.4, "HDI (Income Composition of Resources)": 0.680},
    {"Country": "Republic of Moldova", "Year": 2004, "Avg. Years of Schooling": 10.4, "HDI (Income Composition of Resources)": 0.700},
    {"Country": "Republic of Moldova", "Year": 2005, "Avg. Years of Schooling": 10.5, "HDI (Income Composition of Resources)": 0.710},
    {"Country": "Republic of Moldova", "Year": 2006, "Avg. Years of Schooling": 10.6, "HDI (Income Composition of Resources)": 0.730},
    {"Country": "Republic of Moldova", "Year": 2007, "Avg. Years of Schooling": 10.6, "HDI (Income Composition of Resources)": 0.750},
    {"Country": "Republic of Moldova", "Year": 2008, "Avg. Years of Schooling": 10.7, "HDI (Income Composition of Resources)": 0.770},
    {"Country": "Republic of Moldova", "Year": 2009, "Avg. Years of Schooling": 10.8, "HDI (Income Composition of Resources)": 0.780},
    {"Country": "Republic of Moldova", "Year": 2010, "Avg. Years of Schooling": 10.9, "HDI (Income Composition of Resources)": 0.800},
    {"Country": "Republic of Moldova", "Year": 2011, "Avg. Years of Schooling": 11.0, "HDI (Income Composition of Resources)": 0.820},
    {"Country": "Republic of Moldova", "Year": 2012, "Avg. Years of Schooling": 11.1, "HDI (Income Composition of Resources)": 0.830},
    {"Country": "Republic of Moldova", "Year": 2013, "Avg. Years of Schooling": 11.2, "HDI (Income Composition of Resources)": 0.840},
    {"Country": "Republic of Moldova", "Year": 2014, "Avg. Years of Schooling": 11.3, "HDI (Income Composition of Resources)": 0.850},
    {"Country": "Republic of Moldova", "Year": 2015, "Avg. Years of Schooling": 11.4, "HDI (Income Composition of Resources)": 0.860}
]
uk_data = [
    {'Country': 'United Kingdom', 'Year': 2000, 'Avg. Years of Schooling': 11.9, 'HDI (Income Composition of Resources)': 0.835},
    {'Country': 'United Kingdom', 'Year': 2001, 'Avg. Years of Schooling': 12.0, 'HDI (Income Composition of Resources)': 0.838},
    {'Country': 'United Kingdom', 'Year': 2002, 'Avg. Years of Schooling': 12.1, 'HDI (Income Composition of Resources)': 0.841},
    {'Country': 'United Kingdom', 'Year': 2003, 'Avg. Years of Schooling': 12.2, 'HDI (Income Composition of Resources)': 0.844},
    {'Country': 'United Kingdom', 'Year': 2004, 'Avg. Years of Schooling': 12.3, 'HDI (Income Composition of Resources)': 0.847},
    {'Country': 'United Kingdom', 'Year': 2005, 'Avg. Years of Schooling': 12.4, 'HDI (Income Composition of Resources)': 0.85},
    {'Country': 'United Kingdom', 'Year': 2006, 'Avg. Years of Schooling': 12.6, 'HDI (Income Composition of Resources)': 0.853},
    {'Country': 'United Kingdom', 'Year': 2007, 'Avg. Years of Schooling': 12.7, 'HDI (Income Composition of Resources)': 0.856},
    {'Country': 'United Kingdom', 'Year': 2008, 'Avg. Years of Schooling': 12.8, 'HDI (Income Composition of Resources)': 0.859},
    {'Country': 'United Kingdom', 'Year': 2009, 'Avg. Years of Schooling': 12.9, 'HDI (Income Composition of Resources)': 0.862},
    {'Country': 'United Kingdom', 'Year': 2010, 'Avg. Years of Schooling': 13.0, 'HDI (Income Composition of Resources)': 0.864},
    {'Country': 'United Kingdom', 'Year': 2011, 'Avg. Years of Schooling': 13.1, 'HDI (Income Composition of Resources)': 0.867},
    {'Country': 'United Kingdom', 'Year': 2012, 'Avg. Years of Schooling': 13.2, 'HDI (Income Composition of Resources)': 0.869},
    {'Country': 'United Kingdom', 'Year': 2013, 'Avg. Years of Schooling': 13.3, 'HDI (Income Composition of Resources)': 0.871},
    {'Country': 'United Kingdom', 'Year': 2014, 'Avg. Years of Schooling': 13.4, 'HDI (Income Composition of Resources)': 0.873},
    {'Country': 'United Kingdom', 'Year': 2015, 'Avg. Years of Schooling': 13.5, 'HDI (Income Composition of Resources)': 0.875}
]

tanzania_data = [    {'Country': 'Tanzania', 'Year': 2000, 'Avg. Years of Schooling': 5.3, 'HDI (Income Composition of Resources)': 0.398},    {'Country': 'Tanzania', 'Year': 2001, 'Avg. Years of Schooling': 5.3, 'HDI (Income Composition of Resources)': 0.404},    {'Country': 'Tanzania', 'Year': 2002, 'Avg. Years of Schooling': 5.4, 'HDI (Income Composition of Resources)': 0.411},    {'Country': 'Tanzania', 'Year': 2003, 'Avg. Years of Schooling': 5.4, 'HDI (Income Composition of Resources)': 0.417},    {'Country': 'Tanzania', 'Year': 2004, 'Avg. Years of Schooling': 5.5, 'HDI (Income Composition of Resources)': 0.424},    {'Country': 'Tanzania', 'Year': 2005, 'Avg. Years of Schooling': 5.5, 'HDI (Income Composition of Resources)': 0.431},    {'Country': 'Tanzania', 'Year': 2006, 'Avg. Years of Schooling': 5.6, 'HDI (Income Composition of Resources)': 0.438},    {'Country': 'Tanzania', 'Year': 2007, 'Avg. Years of Schooling': 5.7, 'HDI (Income Composition of Resources)': 0.445},    {'Country': 'Tanzania', 'Year': 2008, 'Avg. Years of Schooling': 5.7, 'HDI (Income Composition of Resources)': 0.452},    {'Country': 'Tanzania', 'Year': 2009, 'Avg. Years of Schooling': 5.8, 'HDI (Income Composition of Resources)': 0.459},    {'Country': 'Tanzania', 'Year': 2010, 'Avg. Years of Schooling': 5.9, 'HDI (Income Composition of Resources)': 0.466},    {'Country': 'Tanzania', 'Year': 2011, 'Avg. Years of Schooling': 5.9, 'HDI (Income Composition of Resources)': 0.473},    {'Country': 'Tanzania', 'Year': 2012, 'Avg. Years of Schooling': 6.0, 'HDI (Income Composition of Resources)': 0.480},    {'Country': 'Tanzania', 'Year': 2013, 'Avg. Years of Schooling': 6.1, 'HDI (Income Composition of Resources)': 0.487},    {'Country': 'Tanzania', 'Year': 2014, 'Avg. Years of Schooling': 6.2, 'HDI (Income Composition of Resources)': 0.494},    {'Country': 'Tanzania', 'Year': 2015, 'Avg. Years of Schooling': 6.3, 'HDI (Income Composition of Resources)': 0.502}]
usa_data = [    {"Country": "USA", "Year": 2000, "Avg. Years of Schooling": 12.7, "HDI (Income Composition of Resources)": 0.854},    {"Country": "USA", "Year": 2001, "Avg. Years of Schooling": 12.8, "HDI (Income Composition of Resources)": 0.859},    {"Country": "USA", "Year": 2002, "Avg. Years of Schooling": 12.8, "HDI (Income Composition of Resources)": 0.864},    {"Country": "USA", "Year": 2003, "Avg. Years of Schooling": 12.9, "HDI (Income Composition of Resources)": 0.869},    {"Country": "USA", "Year": 2004, "Avg. Years of Schooling": 13.0, "HDI (Income Composition of Resources)": 0.874},    {"Country": "USA", "Year": 2005, "Avg. Years of Schooling": 13.1, "HDI (Income Composition of Resources)": 0.878},    {"Country": "USA", "Year": 2006, "Avg. Years of Schooling": 13.2, "HDI (Income Composition of Resources)": 0.882},    {"Country": "USA", "Year": 2007, "Avg. Years of Schooling": 13.3, "HDI (Income Composition of Resources)": 0.884},    {"Country": "USA", "Year": 2008, "Avg. Years of Schooling": 13.4, "HDI (Income Composition of Resources)": 0.883},    {"Country": "USA", "Year": 2009, "Avg. Years of Schooling": 13.4, "HDI (Income Composition of Resources)": 0.879},    {"Country": "USA", "Year": 2010, "Avg. Years of Schooling": 13.5, "HDI (Income Composition of Resources)": 0.879},    {"Country": "USA", "Year": 2011, "Avg. Years of Schooling": 13.5, "HDI (Income Composition of Resources)": 0.883},    {"Country": "USA", "Year": 2012, "Avg. Years of Schooling": 13.6, "HDI (Income Composition of Resources)": 0.886},    {"Country": "USA", "Year": 2013, "Avg. Years of Schooling": 13.7, "HDI (Income Composition of Resources)": 0.888},    {"Country": "USA", "Year": 2014, "Avg. Years of Schooling": 13.7, "HDI (Income Composition of Resources)": 0.890},    {"Country": "USA", "Year": 2015, "Avg. Years of Schooling": 13.8, "HDI (Income Composition of Resources)": 0.892}]


df1 = pd.DataFrame(czechia_data)
df2 = pd.DataFrame(Côte_dIvoire_data)
df3 = pd.DataFrame(congo_data)
df4 = pd.DataFrame(south_korea_data)
df5 = pd.DataFrame(Moldova_data)
df6 = pd.DataFrame(uk_data)
df7 = pd.DataFrame(tanzania_data)
df8 = pd.DataFrame(usa_data)

df = pd.concat([df1, df2, df3, df4, df5, df6, df7, df8], ignore_index=True)

df

df.to_csv('required_data.csv')
